refactor: Rename config and inline_config in x_legacy
Instead of updating these files to make them work with the new rules config forma. We will create new files and use the new one or the legacy one regarding of the specified config
