+++ lib/config_legacy.js
