+++ lib/inline_config_legacy.js
