+++ lib/linter.js
@@ -1,8 +1,8 @@
 const Parser = require("./parser");
 const knife = require("./knife");
 const presets = require("./presets");
-const Config = require("./config");
-const InlineConfig = require("./inline_config");
+const Config = require("./config_legacy");
+const InlineConfig = require("./inline_config_legacy");
 const { flatten } = require("./utils/arrays");
 /**
  * A linter is configured with a set of rules that are fed the raw
